+++ src/core/buffer.ts
@@ -2,8 +2,8 @@
  * Flash Buffer implementation
  *
  * @developer           Elijah Rastorguev
- * @version             1.1.3
- * @build               1012
+ * @version             1.1.5
+ * @build               1014
  * @git                 https://github.com/devsdaddy/flash-buffer/
  * @docs                https://github.com/devsdaddy/flash-buffer/#readme
  * @updated             20.04.2026
@@ -256,6 +256,12 @@ export class FlashBuffer implements Freezable{
     }
 
     /**
+     * Read boolean value
+     * @returns {boolean} Bool value
+     */
+    public readBool() : boolean { return (this.readPrimitive((dv, off) => dv.getUint8(off), 1) !== 0); }
+
+    /**
      * Read Int8
      * @returns {number} Int8 value
      */
@@ -376,6 +382,13 @@ export class FlashBuffer implements Freezable{
         this._offset += byteLength;
         return this;
     }
+
+    /**
+     * Write boolean value
+     * @param value {boolean} Bool value
+     * @returns {FlashBuffer} current buffer instance
+     */
+    public writeBool(value : boolean) : this { return this.writePrimitive((dv, off, v) => dv.setInt8(off, v), (value) ? 1 : 0, 1); }
 
     /**
      * Write Int8 value