+++ src/schema/decorators.ts
@@ -2,11 +2,11 @@
  * Flash Buffer Decorators
  *
  * @developer           Elijah Rastorguev
- * @version             1.1.0
- * @build               1005
+ * @version             1.1.5
+ * @build               1007
  * @git                 https://github.com/devsdaddy/flash-buffer/
  * @docs                https://github.com/devsdaddy/flash-buffer/#readme
- * @updated             13.04.2026
+ * @updated             20.04.2026
  */
 import 'reflect-metadata';
 import {ProtobufType} from "../protobuf";
@@ -17,7 +17,8 @@ const SCHEMA_METADATA_KEY = Symbol('binary:schema');
 /* Flash buffer field types */
 export type FlashBufferType = 'uint8' | 'int8' | 'uint16' | 'int16' | 'uint32' | 'int32' | 'uint64' |
     'int64' | 'float32' | 'float64' | 'string' | 'cstring' | 'varuint' | 'varint' |
-    'varuint64' | 'sint32' | 'sint64' | 'fixed32' | 'fixed64' | 'sfixed32' | 'sfixed64';
+    'varuint64' | 'sint32' | 'sint64' | 'fixed32' | 'fixed64' | 'sfixed32' | 'sfixed64' |
+    'bool' | { array: FlashBufferType } | { object: Record<string, FlashBufferType> } | { nullable: FlashBufferType };
 
 /**
  * Field Options
@@ -26,6 +27,7 @@ export interface FieldOptions {
     type?: FlashBufferType;
     littleEndian?: boolean;
     length?: number; // for strings: max length, or fixed length
+    optional?: boolean;
 
     // For protobuf
     protoType?: ProtobufType;