+++ src/schema/schema.ts
@@ -2,16 +2,15 @@
  * Flash Buffer schema support
  *
  * @developer           Elijah Rastorguev
- * @version             1.1.0
- * @build               1005
+ * @version             1.1.5
+ * @build               1007
  * @git                 https://github.com/devsdaddy/flash-buffer/
  * @docs                https://github.com/devsdaddy/flash-buffer/#readme
- * @updated             13.04.2026
+ * @updated             20.04.2026
  */
 /* Import required modules */
 import {FlashBuffer} from "../core/buffer";
 import {FieldOptions, getSchemaFields} from "./decorators";
-import {ProtobufReader, ProtobufWriter} from "../protobuf";
 
 /**
  * Flash buffer schema
@@ -50,6 +49,7 @@ export class FlashBufferSchema {
                 case 'sfixed64': buf.writeSFixed64(value); break;
                 case 'string': buf.writeString(value, 'utf-8', true); break;
                 case 'cstring': buf.writeCString(value); break;
+                case 'bool': buf.writeBool(value); break;
                 default: throw new Error(`Unsupported type: ${options.type}`);
             }
         }
@@ -92,6 +92,7 @@ export class FlashBufferSchema {
                     break;
                 }
                 case 'cstring': value = buffer.readCString(); break;
+                case 'bool': value = buffer.readBool(); break;
                 default: throw new Error(`Unsupported type: ${options.type}`);
             }
             (obj as any)[propertyKey] = value;